Initial this repo
Create the files and just leave them blank.

This repo is an enhancement of Scrapy cookies middleware.

The original one saves cookies in a defaultdict with the default value
CookieJar. As the enhancement, this repo would like to follow these
principles:

- Keep the behavior same as the original one under all settings
  supported by the original one
- Pass all test cases from the original one
- Write test cases for all features privoded in this enhancement
- Support all Python versions which Scrapy supports
- Fully documented following the structure and the style of Scrapy

And the most important, this enhancement will privode the following
features:

- Design a backend system for this middleware to save and read cookies
- Privoded several widely-used backends