Extract cookies middleware from Scrapy
Extract all cookies middleware related codes, including test cases,
travis and codecov configurations, etc.

In this commit, the purpose is to exrtact the minimum parts from Scrapy
to develop cookies middleware.

Althrough the best effort is taken to keep the code same as the
original one, some changes are made to keep the consistent between the
code and this repo:

- MANIFEST.in
- setup.py

while some other changes are made to pass the test cases, and also
the CI tool (Travis here) and Codecov:

- .coveragerc
- .travis.yml
- tox.ini

This repo maybe need further tuning to fit this repo.