Save the value (jar) back to the storage
Add one line of code to save the value (jar) back to the storage by
key.

In the original the storage is a defaultdict, the change of value
(jar) will save directly (the value is mutable object and it is
referred by address in dict).

In the future the storage could be various types, the saving of values
have to be explicit.