Add two methods to the storage object
In this commit, two methods are added to the storage object:

- open_spider
- close_spider

These two methods intend to be called in the middleware when signals
spider_opened and spider_closed are sent, to initialize and finalize
the storage object if it is necessary.

In the same time, the related code is added in the test cases.